+++ src/types/client.ts
@@ -1,16 +1,34 @@
-import { ApolloClient } from "apollo-client";
-import { InMemoryCache } from "apollo-cache-inmemory";
-import { ApolloLink } from "apollo-link";
+import { DocumentNode, ExecutionResult } from "graphql";
 import { Adapter } from "../adapters/Adapter";
+import { FetchPolicy, ErrorPolicy } from "./query";
 
 export interface ClientConfig {
   name?: string;
-  apollo?: ApolloClient<any>;
-  url?: string;
-  credentials?: string;
+  graphQlClient: GraphQlClient;
   adapter?: Adapter;
-  link?: ApolloLink;
-  cache?: InMemoryCache;
-  useGETForQueries?: boolean;
-  debugMode?: boolean;
+}
+
+type OperationVariables = {
+  [key: string]: any;
+};
+
+export interface QueryOptions {
+  query: DocumentNode;
+  variables?: OperationVariables;
+  errorPolicy?: ErrorPolicy;
+  fetchResults?: boolean;
+  context?: any;
+  fetchPolicy?: FetchPolicy;
+}
+
+export interface MutationOptions {
+  mutation: DocumentNode;
+  variables?: OperationVariables;
+  context?: any;
+  fetchPolicy?: FetchPolicy;
+}
+
+export interface GraphQlClient {
+  query(options: QueryOptions): Promise<ExecutionResult>;
+  mutate(options: MutationOptions): Promise<ExecutionResult>;
 }